\name{S4Vectors-internals}
\alias{S4Vectors-internals}

\alias{I}
\alias{fold}
\alias{isSingleInteger}
\alias{isSingleNumber}
\alias{isSingleNumberOrNA}
\alias{isSingleString}
\alias{isSingleStringOrNA}
\alias{isTRUEorFALSE}
\alias{new2}
\alias{recycleArg}
\alias{recycleCharacterArg}
\alias{recycleIntegerArg}
\alias{recycleLogicalArg}
\alias{recycleNumericArg}
\alias{setMethods}
\alias{setValidity2}
\alias{wmsg}

\alias{class:atomic}
\alias{atomic-class}
\alias{atomic}
\alias{class:character_OR_NULL}
\alias{character_OR_NULL-class}
\alias{character_OR_NULL}
\alias{class:DataFrame_OR_NULL}
\alias{DataFrame_OR_NULL-class}
\alias{DataFrame_OR_NULL}
\alias{class:vector_OR_factor}
\alias{vector_OR_factor-class}
\alias{vector_OR_factor}

\alias{coerce,ANY,AsIs-method}
\alias{coerce,ANY,vector-method}

\title{Low-level internal helpers}

\description{
  Low-level helper functions, coercions, and union classes used internally
  by the \pkg{S4Vectors} package.
}


\keyword{internal}
\keyword{classes}
\keyword{methods}
