\name{show-utils}

\alias{show-utils}

\alias{coolcat}

\alias{classNameForDisplay}
\alias{classNameForDisplay,ANY-method}
\alias{classNameForDisplay,AsIs-method}
\alias{showAsCell}
\alias{showAsCell,ANY-method}
\alias{showAsCell,numeric-method}
\alias{showAsCell,AsIs-method}
\alias{showAsCell,list-method}
\alias{showAsCell,data.frame-method}
\alias{makeNakedCharacterMatrixForDisplay}
\alias{makeNakedCharacterMatrixForDisplay,ANY-method}
\alias{cbind_mcols_for_display}
\alias{makePrettyMatrixForCompactPrinting}
\alias{makeClassinfoRowForCompactPrinting}

\title{Display utilities}

\description{
  Low-level utility functions and classes defined in the \pkg{S4Vectors}
  package to support display of vector-like objects. They are not
  intended to be used directly.
}

\keyword{utilities}
\keyword{classes}
\keyword{methods}
