\name{S4Vectors internals}

% Stuff from R/S4-utils.R:
\alias{class:characterORNULL}
\alias{characterORNULL-class}
\alias{characterORNULL}
\alias{class:vectorORfactor}
\alias{vectorORfactor-class}
\alias{vectorORfactor}
\alias{coerce,ANY,vector-method}
\alias{setValidity2}
\alias{new2}
\alias{setMethods}

% Stuff from R/utils.R:
\alias{wmsg}
\alias{.Call2}
\alias{get_showHeadLines}
\alias{get_showTailLines}
\alias{printAtomicVectorInAGrid}

% Low-level helper functions from R/normarg-utils.R:
\alias{isTRUEorFALSE}
\alias{isSingleInteger}
\alias{isSingleNumber}
\alias{isSingleString}
\alias{isSingleNumberOrNA}
\alias{isSingleStringOrNA}
\alias{recycleIntegerArg}
\alias{recycleNumericArg}
\alias{fold}

\title{S4Vectors internals}

\description{
  Objects, classes and methods defined in the \pkg{S4Vectors} package
  that are not intended to be used directly.
}

\keyword{internal}
\keyword{classes}
\keyword{methods}
