\name{show-utils}

\alias{show-utils}

\alias{classNameForDisplay}
\alias{classNameForDisplay,ANY-method}
\alias{classNameForDisplay,AsIs-method}
\alias{showAsCell}
\alias{showAsCell,ANY-method}
\alias{showAsCell,AsIs-method}
\alias{showAsCell,Date-method}
\alias{showAsCell,POSIXt-method}

\title{Display utilities}

\description{
  Low-level utility functions and classes defined in the \pkg{S4Vectors}
  package to support display of vector-like objects. They are not
  intended to be used directly.
}

\keyword{utilities}
\keyword{classes}
\keyword{methods}
