\name{subsetting-utils}

\alias{subsetting-utils}

\alias{class:NSBS}
\alias{NSBS-class}

\alias{length,NSBS-method}
\alias{anyDuplicated.NSBS}
\alias{anyDuplicated,NSBS-method}
\alias{isStrictlySorted,NSBS-method}

\alias{NSBS}
\alias{NSBS,NSBS-method}
\alias{NSBS,missing-method}
\alias{NSBS,NULL-method}
\alias{NSBS,numeric-method}
\alias{NSBS,logical-method}
\alias{NSBS,character-method}
\alias{NSBS,factor-method}
\alias{NSBS,array-method}

\alias{as.integer,NativeNSBS-method}

\alias{as.integer,RangeNSBS-method}
\alias{length,RangeNSBS-method}
\alias{anyDuplicated,RangeNSBS-method}
\alias{isStrictlySorted,RangeNSBS-method}
\alias{show,RangeNSBS-method}

\alias{normalizeSingleBracketSubscript}
\alias{normalizeSingleBracketReplacementValue}
\alias{normalizeSingleBracketReplacementValue,ANY-method}

\alias{extractROWS}
\alias{extractROWS,ANY,ANY-method}
\alias{extractROWS,vectorORfactor,RangeNSBS-method}
\alias{replaceROWS}
\alias{replaceROWS,ANY-method}

\alias{normalizeDoubleBracketSubscript}

\alias{getListElement}
\alias{setListElement}


\title{Subsetting utilities}

\description{
  Low-level utility functions and classes defined in the \pkg{S4Vectors}
  package to support subsetting of vector-like objects. They are not
  intended to be used directly.
}

\keyword{utilities}
\keyword{classes}
\keyword{methods}
